fix: align urc alpha branch with current sdk stack
