+++ Packages/com.nekometer.esnya.udon-radio-communications-sf/Samples~/SaccFlightSamples/Demo.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f995de9df608cef4cac3c505be1f0eff
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
