+++ Packages/com.nekometer.esnya.udon-radio-communications-sf/Samples~/SaccFlightSamples/SampleVehicles.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e5563a9f73981af4aa2a6d5b93ec4f19
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
