+++ Packages/com.nekometer.esnya.udon-radio-communications-sf/Samples~/SaccFlightSamples/SampleVehicles/SF-1Main+COM.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5aa0696f016f7214b9ee757e6f549ee6
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
