+++ Packages/com.nekometer.esnya.udon-radio-communications-sf/Samples~/SaccFlightSamples/SampleVehicles/SH-1Main+COM.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a35c399fcf052ad42aa9da31a8cfd147
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
