+++ Packages/com.nekometer.esnya.udon-radio-communications-sf/package.json
@@ -3,18 +3,21 @@
   "displayName": "Udon Radio Communications SaccFlight Integration",
   "version": "1.0.0-alpha.1",
   "gitDependencies": {
-    "com.merlinvr.udonsharp": "https://github.com/vrchat-community/UdonSharp.git?path=/Packages/com.vrchat.UdonSharp#1.0.1",
-    "com.vrchat.base": "https://github.com/vrchat/packages.git?path=/packages/com.vrchat.base#3.0.4",
-    "com.vrchat.worlds": "https://github.com/vrchat/packages.git?path=/packages/com.vrchat.worlds#3.0.4",
-    "com.nekometer.esnya.udon-rado-communications": "https://github.com/esnya/UdonRadioCommunications.git?path=Packages/com.nekometer.esnya.udon-radio-communications"
+    "com.vrchat.base": "https://github.com/vrchat/packages.git?path=/packages/com.vrchat.base#3.10.2",
+    "com.vrchat.worlds": "https://github.com/vrchat/packages.git?path=/packages/com.vrchat.worlds#3.10.2",
+    "com.nekometer.esnya.udon-radio-communications": "https://github.com/esnya/UdonRadioCommunications.git?path=Packages/com.nekometer.esnya.udon-radio-communications#alpha"
+  },
+  "vpmDependencies": {
+    "com.vrchat.worlds": "3.10.2",
+    "com.nekometer.esnya.udon-radio-communications": ">=5.0.0-alpha"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/esnya/UdonRadioCommunication.git"
+    "url": "git+https://github.com/esnya/UdonRadioCommunications.git"
   },
   "author": {
     "name": "esnya"
   },
   "license": "MIT",
-  "unity": "2019.4"
+  "unity": "2022.3"
 }
