+++ Packages/com.nekometer.esnya.udon-radio-communications/Scripts/UdonRadioCommunication.Runtime.asmdef
@@ -3,7 +3,7 @@
     "references": [
         "VRC.Udon",
         "UdonSharp.Runtime",
-        "UdonSharp.Editor",
+        "VRC.SDKBase",
         "Unity.TextMeshPro"
     ],
     "includePlatforms": [],
